Working context for light dark mode.
